fix(protocol-designer): fix settings page layout hen devf f is on
fix settings page layout hen devf ff is on + refactor Settings components

close AUTH-1545 and AUTH-
