+++ protocol-designer/src/ProtocolRoutes.tsx
@@ -73,7 +73,7 @@ export function ProtocolRoutes(): JSX.Element {
     >
       <Navigation />
       <Kitchen>
-        <Box width="100%" height="100%">
+        <Box width="100%" height="100%" id="protocol-routes">
           <GateModal />
           <LabwareUploadModal />
           <FileUploadMessagesModal />
